+++ Sources/MapboxSearch/PublicAPI/Use Cases/Address Autofill/Models/Suggestion/AddressAutofill.Suggestion+SearchResult.swift
@@ -27,7 +27,8 @@ extension AddressAutofill.Suggestion {
             name: searchResult.name,
             formattedAddress: formattedAddress,
             coordinate: searchResult.coordinate,
-            addressComponents: try address.toAutofillComponents()
+            addressComponents: try address.toAutofillComponents(),
+            underlying: .result(searchResult)
         )
     }
 }