+++ Sources/MapboxSearch/PublicAPI/Use Cases/Place Autocomplete/Models/PlaceAutocomplete+Suggestion.swift
@@ -12,8 +12,8 @@ public extension PlaceAutocomplete {
         public let description: String?
 
         /// Geographic point.
-        public let coordinate: CLLocationCoordinate2D
-        
+        public let coordinate: CLLocationCoordinate2D?
+
         /// Icon name according to [Mapbox Maki icon set](https://github.com/mapbox/maki/)
         public let iconName: String?
 
@@ -32,12 +32,14 @@ public extension PlaceAutocomplete {
         /// List of points near `coordinate`, that represents entries to associated building.
         public let routablePoints: [RoutablePoint]
         
+        /// Underlying data provided by core SDK and API used to construct this Suggestion instance.
+        /// Useful for any follow-up API calls or unit test validation.
         let underlying: Underlying
 
         init(
             name: String,
             description: String?,
-            coordinate: CLLocationCoordinate2D,
+            coordinate: CLLocationCoordinate2D?,
             iconName: String?,
             distance: CLLocationDistance?,
             estimatedTime: Measurement<UnitDuration>?,